- Setup a sqlite3 database so messages can persist
- Updated the server to utilize the database.
  NOTE: use the route "localhost:5000/erase" to erase the database for 
testing purposes. This is so we don't have to manually erase everything.

- Updated client fake messages to be randomized string of a random 
length.

- Tested TLS successfully.
  NOTE: TLS is possible, and I did have it working, however since we 
don't have real SSL certificates we get a giant warning page before we 
connect. During development we should work without TLS actually enabled, 
under the assumption that we could always enable it if this was ever 
released.